+++ pkg/adapter/runtime_remote_supported.go
@@ -1,0 +1,1 @@
+package adapter
